+++ src/rp2_common/hardware_xosc/include/hardware/xosc.h
@@ -10,6 +10,15 @@
 #include "pico.h"
 #include "hardware/structs/xosc.h"
 
+
+// Allow lengthening startup delay to accommodate slow-starting oscillators
+
+// PICO_CONFIG: PICO_XOSC_STARTUP_DELAY_MULTIPLIER, Multiplier to lengthen xosc startup delay to accommodate slow-starting oscillators, type=int, min=1, default=1, group=hardware_xosc
+#ifndef PICO_XOSC_STARTUP_DELAY_MULTIPLIER
+#define PICO_XOSC_STARTUP_DELAY_MULTIPLIER 1
+#endif
+
+
 #ifdef __cplusplus
 extern "C" {
 #endif
